assist-core/src/main/java/me/dslztx/assist/util/filetype/PSDMatcher.java
package me.dslztx.assist.util.filetype;

import me.dslztx.assist.bean.filetype.FileType;
import me.dslztx.assist.util.ArrayAssist;

public class PSDMatcher implements FileTypeMatcher {

    @Override
    public boolean match(byte[] data) {
        if (ArrayAssist.isEmpty(data)) {
            return false;
        }

        if (data.length >= 4) {
            if (data[0] == (byte)0x38 && data[1] == (byte)0x42 && data[2] == (byte)0x50 && data[3] == (byte)0x53) {
                return true;
            }
        }

        return false;
    }

    @Override
    public FileType obtainFileType() {
        return FileType.PSD;
    }
}
